@@ -11,67 +11,65 @@
     "dev:mock": "cross-env MOCK=true npm run dev",
     "build": "npm run clean && cross-env NODE_ENV=production webpack --mode=production",
     "analyze": "cross-env NODE_ENV=production ANALYZE=true webpack --mode=production",
+    "lint": "eslint . --fix",
     "pretty": "prettier --write './**/*.(js|ts|jsx|tsx|md|html)'",
     "pack": "rimraf figma-plugin.zip && npm run build && cd dist && bestzip ../figma-plugin.zip *",
-    "prepare": "husky install"
+    "prepare": "husky"
   },
   "author": "",
   "license": "",
   "devDependencies": {
-    "@babel/plugin-transform-runtime": "~7.15.0",
-    "@figma/plugin-typings": "^1.28.0",
-    "@types/lodash": "~4.14.172",
-    "@types/react": "~17.0.15",
-    "@types/react-dom": "~17.0.9",
-    "@types/uuid": "~8.3.1",
-    "@typescript-eslint/eslint-plugin": "~4.29.0",
-    "@typescript-eslint/parser": "~4.29.0",
-    "babel-loader": "~8.2.2",
-    "babel-plugin-import": "~1.13.3",
-    "babel-plugin-named-asset-import": "~0.3.7",
-    "babel-preset-react-app": "~10.0.0",
-    "bestzip": "~2.2.0",
-    "cross-env": "~7.0.3",
-    "css-loader": "^6.2.0",
-    "eslint": "~7.32.0",
-    "eslint-config-prettier": "~8.3.0",
-    "eslint-config-standard": "~16.0.2",
-    "eslint-plugin-import": "~2.23.4",
-    "eslint-plugin-node": "~11.1.0",
-    "eslint-plugin-promise": "~5.1.0",
-    "eslint-plugin-react": "~7.24.0",
-    "eslint-plugin-react-hooks": "~4.2.0",
-    "eslint-webpack-plugin": "~3.0.1",
-    "html-webpack-plugin": "^5.3.2",
-    "husky": "~7.0.2",
-    "less": "^4.1.1",
-    "less-loader": "^10.0.1",
-    "lint-staged": "~11.1.2",
-    "npm-run-all": "~4.1.5",
-    "postcss": "~8.3.6",
+    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@eslint/js": "~9.39.1",
+    "@figma/plugin-typings": "^1.103.0",
+    "@types/lodash": "^4.17.12",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/uuid": "^10.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
+    "babel-loader": "^9.2.1",
+    "babel-plugin-named-asset-import": "^0.3.8",
+    "babel-preset-react-app": "^10.0.1",
+    "bestzip": "^2.2.1",
+    "cross-env": "^7.0.3",
+    "css-loader": "^7.1.2",
+    "eslint": "^9.15.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-promise": "^7.1.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "html-webpack-plugin": "^5.6.3",
+    "husky": "^9.1.7",
+    "less": "^4.2.0",
+    "less-loader": "^12.2.0",
+    "lint-staged": "^15.2.10",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.5.1",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^6.1.1",
-    "postcss-preset-env": "~7.6.0",
-    "prettier": "^2.3.2",
-    "react-dev-utils": "~11.0.4",
-    "style-loader": "^3.2.1",
-    "typescript": "^4.3.5",
+    "postcss-loader": "^8.1.1",
+    "postcss-preset-env": "^10.1.1",
+    "prettier": "^3.3.3",
+    "react-dev-utils": "^12.0.1",
+    "react-refresh": "^0.14.2",
+    "rimraf": "^6.0.1",
+    "style-loader": "^4.0.0",
+    "typescript": "^5.7.2",
     "url-loader": "^4.1.1",
-    "webpack": "~5.48.0",
-    "webpack-bundle-analyzer": "~4.4.2",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "~4.1.0"
+    "webpack": "^5.96.1",
+    "webpack-bundle-analyzer": "^4.10.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0"
   },
   "dependencies": {
-    "@hot-loader/react-dom": "~17.0.1",
-    "antd": "~4.20.5",
-    "axios": "~0.21.4",
-    "eventemitter3": "~4.0.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hot-loader": "~4.13.0",
-    "react-router-dom": "~5.3.0",
-    "uuid": "~8.3.2"
+    "antd": "^5.22.2",
+    "axios": "^1.7.9",
+    "eventemitter3": "^5.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
+    "uuid": "^11.0.3"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": "eslint --cache --fix",
