@@ -3,12 +3,15 @@
  */
 
 import React from 'react'
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
+import 'antd/dist/reset.css'
 import App from 'pages'
 import EventEmitter from './utils/event'
 
 // 渲染
-render(<App />, document.getElementById('my-plugin'))
+const container = document.getElementById('my-plugin')!
+const root = createRoot(container)
+root.render(<App />)
 
 // 通过 event emitter 处理消息，需要访问来自figma的消息时，通过如下方式
 // import EventEmitter from 'utils/event'
@@ -26,7 +29,8 @@ onmessage = (evt) => {
 // @ts-ignore
 if (module.hot) {
   // @ts-ignore
-  module.hot.accept('./ui.tsx', () => {
-    render(<App />, document.getElementById('my-plugin'))
+  module.hot.accept('./pages', () => {
+    // Re-render the app when pages change
+    root.render(<App />)
   })
 }
